requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
python-decouple==3.8
Pillow==11.3.0
django-filter==25.1
drf-yasg==1.21.10